venv_requirements.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.75.0
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.0.5
attrs==25.4.0
autoflake==2.3.1
babel==2.17.0
backrefs==6.1
bandit==1.9.2
beautifulsoup4==4.14.3
black==25.12.0
bleach==6.3.0
boolean.py==5.0
CacheControl==0.14.4
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
coverage==7.13.0
cryptography==46.0.3
cuda-bindings==12.9.4
cuda-core==0.3.2
cuda-pathfinder==1.3.3
cudensitymat-cu12==0.3.2
cupauliprop-cu12==0.1.0
cupy-cuda12x==13.6.0
cuquantum-cu12==25.11.0
cuquantum-python==25.11.1
cuquantum-python-cu12==25.11.1
custabilizer-cu12==0.1.0
custatevec-cu12==1.11.0
cutensor-cu12==2.4.1
cutensornet-cu12==2.10.1
cycler==0.12.1
cyclonedx-python-lib==11.6.0
dataclasses-json==0.6.7
dbus-python==1.4.0
debugpy==1.8.19
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
dill==0.4.0
distlib==0.4.0
distro==1.9.0
docplex==2.30.251
docstring_parser==0.17.0
execnet==2.1.2
executing==2.2.1
fakeredis==2.32.1
falcon==4.2.0
fastapi==0.124.4
fastjsonschema==2.21.2
fastrlock==0.8.3
filelock==3.20.1
flake8==7.3.0
fonttools==4.61.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.12.0
ghp-import==2.1.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
gudhi==3.11.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
hyperframe==6.1.0
ibm-cloud-sdk-core==3.24.2
ibm-platform-services==0.72.0
identify==2.6.15
idna==3.11
importlib_metadata==8.7.0
iniconfig==2.3.0
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
isort==7.0.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
json5==0.12.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.7.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
langchain==1.2.0
langchain-anthropic==1.3.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.2.5
langchain-ollama==1.0.1
langchain-openai==1.1.3
langchain-text-splitters==1.1.0
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.0
langsmith==0.4.59
lark==1.3.1
librt==0.7.4
license-expression==30.4.4
line_profiler==5.0.0
mando==0.7.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.8
matplotlib-inline==0.2.1
mccabe==0.7.0
mdurl==0.1.2
mergedeep==1.3.4
mistune==3.1.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.7.0
mkdocs-material-extensions==1.3.1
mmh3==5.2.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
mypy==1.19.1
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
nodeenv==1.9.1
notebook==7.5.1
notebook_shim==0.2.4
numpy==2.3.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.1.105
nvmath-python==0.6.0
ollama==0.6.1
openai==2.12.0
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.5
ormsgpack==1.12.1
packageurl-python==0.17.6
packaging==25.0
paginate==0.5.7
pandas==2.3.3
pandas-stubs==2.3.3.251201
pandocfilters==1.5.1
parso==0.8.5
pathspec==0.12.1
patsy==1.0.2
pexpect==4.9.0
pillow==11.3.0
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.10.0
platformdirs==4.5.1
playwright==1.57.0
pluggy==1.6.0
portalocker==3.2.0
postgrest==2.27.0
POT==0.9.6.post1
pre_commit==4.5.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.2
psutil==7.1.3
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
py-serializable==2.1.0
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydub==0.25.1
pyee==13.0.0
pyflakes==3.4.0
Pygments==2.19.2
pyiceberg==0.10.0
pyitlib==0.3.1
PyJWT==2.10.1
pymdown-extensions==10.19.1
pyparsing==3.2.5
pyroaring==1.0.3
pyspnego==0.12.0
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-mock==3.15.1
pytest-timeout==2.4.0
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.20
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
pyyaml_env_tag==1.1
pyzmq==27.1.0
qdrant-client==1.16.2
qiskit==2.2.3
qiskit-aer==0.17.2
qiskit-aer-gpu==0.15.1
qiskit-algorithms==0.4.0
qiskit-ibm-runtime==0.44.0
qiskit-optimization==0.7.0
radon==6.0.1
realtime==2.27.0
redis==7.1.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
requests_ntlm==1.3.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rpds-py==0.30.0
ruff==0.14.9
rustworkx==0.17.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==5.2.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
snakeviz==2.2.2
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.3
soupsieve==2.8
SQLAlchemy==2.0.45
stack-data==0.6.3
starlette==0.50.0
statsmodels==0.14.6
stevedore==5.6.0
storage3==2.27.0
StrEnum==0.4.15
strictyaml==1.7.3
structlog==25.5.0
supabase==2.27.0
supabase-auth==2.27.0
supabase-functions==2.27.0
symengine==0.13.0
sympy==1.13.1
tenacity==9.1.2
tenseal==0.3.16
terminado==0.18.1
threadpoolctl==3.6.0
tiktoken==0.12.0
tinycss2==1.4.0
tokenizers==0.22.1
tomli==2.3.0
tomli_w==1.2.0
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.3
triton==3.1.0
typer-slim==0.20.0
types-psutil==7.1.3.20251211
types-pytz==2025.2.0.20251108
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20250913
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uri-template==1.3.0
urllib3==2.6.2
uuid_utils==0.12.0
uvicorn==0.38.0
virtualenv==20.35.4
watchdog==6.0.0
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
wheel==0.45.1
widgetsnbextension==4.0.15
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
